requirements.txt
aiohttp==3.10.10
click==8.1.7
discord.py==2.4.0
groq==0.16.0
motor==3.7.0
Pillow==11.1.0
python-dotenv==1.0.1
Requests==2.32.3
wavelink==3.4.1
